pyproject.toml
# package build
# https://peps.python.org/pep-0517/
[build-system]
requires = ["setuptools>=65", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

# package basics
# https://peps.python.org/pep-0621/
[project]
name = "repo-statistics"
description = "Calculate metrics and statistics for source-code repositories."
keywords = []
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MPLv2 License" }
authors = [
  { email = "evamaxfieldbrown@gmail.com", name = "Eva Maxfield Brown" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dynamic = ["version"]
dependencies = [
  "backoff>=2,<3",
  "dataclasses-json>=0.6,<1",
  "gitpython>=3,<4",
  "gh-tokens-loader>=0.1.2,<1",
  "ghapi>=1,<2",
  "numpy>=1,<2",
  "scipy>=1,<2",
  "polars>=1,<2",
  "pygount>=3,<4",
  "pyyaml>=6,<7",
  "timeout-function-decorator>=2,<3",
  "tqdm>=4,<5",
  "multimetric>=2,<3",
]

# extra dependencies
# https://peps.python.org/pep-0621/#dependencies-optional-dependencies
[project.optional-dependencies]
lint = [
  "ruff>=0.12",
  "ty>=0.0.12",
]
dev = [
  "ipython",
  "jupyterlab",
  "pytest>=5.4.3",
]
coiled = [
  "coiled>=1,<2",
]

[project.urls]
Homepage = "https://github.com/evamaxfield/repo-statistics"
"Bug Tracker" = "https://github.com/evamaxfield/repo-statistics/issues"
Documentation = "https://evamaxfield.github.io/repo-statistics"
"User Support" = "https://github.com/evamaxfield/repo-statistics/issues"

# entry points
# example
# rs-graph-prelinked-dataset-ingestion = "rs_graph.bin.prelinked_dataset_ingestion:main"
# https://peps.python.org/pep-0621/#entry-points
# [project.entry-points."console_scripts"]

# build settings
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["*docs/*", "*tests/*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "py.typed"]

# tools
# https://github.com/charliermarsh/ruff
[tool.ruff]
line-length = 96
target-version = "py312"
lint.extend-select = [
    "E",      # style errors
    "F",      # flakes
    "D",      # pydocstyle
    "I001",   # isort
    "UP",     # pyupgrade
    "N",      # pep8-naming
    # "S",    # bandit
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "A001",   # flake8-builtins
    "RUF",    # ruff-specific rules
    "RUF100", # Unused noqa directive
]
lint.extend-ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D103", # Missing docstring in public function
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D205", # 1 blank line required between summary line and description
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line
    "D413", # Missing blank line after last section
    "D416", # Section name should end with a colon
]

[tool.ruff.lint.per-file-ignores]
"**/notebooks/*" = ["E501", "B018"]
"**/tests/*.py" = ["D"]
"**/migrations/*.py" = ["D"]

[tool.ty.src]
exclude = [".github", ".pre-commit-config.yaml", "pyproject.toml"]